+++ src/platform/aws/src/s3/mod.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod bucket;
 mod error;
 mod object;
 mod storage;
+mod uploader;
 
 pub use self::storage::S3Storage;
